refactor: 将跨模块共享类型移入 closeclaw-common
将以下类型从原模块提取到 crates/common/src/，原位置改为 pub use re-export：

- BootstrapMode → common/src/bootstrap.rs
- CompactConfig → common/src/compaction.rs
- AgentRole, ReasoningLevel → common/src/session_types.rs
- CommunicationConfig, CommunicationCheckResult, CommunicationError,
  check_communication_allowed → common/src/communication.rs
- AgentConfig, SubagentsConfig, AgentPermissions, PermissionLimits,
  ActionPermission, MemoryConfig, ActiveSearcherOverride → common/src/agent_config.rs

原模块通过 pub use closeclaw_common::X 保持对外 API 完全不变。

Source: issue #1252
Type: refactor
