+++ crates/common/Cargo.toml
@@ -8,3 +8,4 @@ license = "MIT"
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+thiserror = "2.0"
