+++ crates/common/src/lib.rs
@@ -1,2 +1,18 @@
+pub mod agent_config;
+pub mod bootstrap;
+pub mod communication;
+pub mod compaction;
+pub mod session_types;
 pub mod verbosity;
+
+pub use agent_config::{
+    ActionPermission, ActiveSearcherOverride, AgentConfig, AgentPermissions, MemoryConfig,
+    PermissionLimits, SubagentsConfig,
+};
+pub use bootstrap::BootstrapMode;
+pub use communication::{
+    check_communication_allowed, CommunicationCheckResult, CommunicationConfig, CommunicationError,
+};
+pub use compaction::CompactConfig;
+pub use session_types::{AgentRole, ReasoningLevel};
 pub use verbosity::VerbosityLevel;
