+++ src/agent/config/config_intersect_tests.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use super::*;
 
 // =====================================================================
